Migrate RunConfig to Plugin with new API
